+++ core/src/builtin_widgets/clip.rs
@@ -6,7 +6,7 @@ pub struct Clip {
 }
 
 impl Render for Clip {
-  fn only_sized_by_parent(&self) -> bool { true }
+  fn size_affected_by_child(&self) -> bool { false }
 
   fn perform_layout(&self, clamp: BoxClamp, ctx: &mut LayoutCtx) -> Size {
     ctx.assert_perform_single_child_layout(clamp);