+++ core/src/test_helper.rs
@@ -248,7 +248,7 @@ impl Render for MockBox {
     size
   }
   #[inline]
-  fn only_sized_by_parent(&self) -> bool { true }
+  fn size_affected_by_child(&self) -> bool { false }
 
   #[inline]
   fn paint(&self, _: &mut PaintingCtx) {}