+++ widgets/src/layout/no_affected_parent_size.rs
@@ -1,32 +1,31 @@
 use ribir_core::prelude::*;
 
-/// OnlySizedByParent implies that the parent is the only input into determining
-/// the widget's size, so layout changes to the subtree do not trigger a parent
-/// relayout.
+/// NoAffectedParentSize implies that the layout changes to the subtree do not
+/// trigger a parent relayout.
 #[derive(SingleChild, Declare)]
-pub struct OnlySizedByParent {}
+pub struct NoAffectedParentSize {}
 
-// `OnlySizedByParent` must be an independent node in the widget tree.
+// `NoAffectedParentSize` must be an independent node in the widget tree.
 // Therefore, any modifications to its child should terminate at
-// `OnlySizedByParent`. Otherwise, if its host is dirty, it implies that the
-// `OnlySizedByParent` node is also dirty, and its parent must be marked as
-// dirty. For instance, if `w2` in a Row[w1, OnlySizedByParent<w2>] is dirty,
-// the Row requires a relayout.
-impl Render for OnlySizedByParent {
+// `NoAffectedParentSize`. Otherwise, if its host is dirty, it implies that the
+// `NoAffectedParentSize` node is also dirty, and its parent must be marked as
+// dirty. For instance, if `w2` in a Row[w1, NoAffectedParentSize<w2>] is
+// dirty, the Row requires a relayout.
+impl Render for NoAffectedParentSize {
   fn perform_layout(&self, clamp: BoxClamp, ctx: &mut LayoutCtx) -> Size {
     ctx
       .perform_single_child_layout(clamp)
       .unwrap_or(ZERO_SIZE)
   }
 
-  fn only_sized_by_parent(&self) -> bool { true }
+  fn size_affected_by_child(&self) -> bool { false }
 }
 
 #[cfg(test)]
 mod tests {
   use ribir_core::{prelude::*, reset_test_env, test_helper::*};
 
-  use crate::layout::OnlySizedByParent;
+  use crate::layout::NoAffectedParentSize;
 
   #[test]
   fn ignore_layout_changed() {
@@ -43,7 +42,7 @@ mod tests {
           *$child1_layout_writer.write() += 1;
         },
       };
-      let child2 = @OnlySizedByParent {
+      let child2 = @NoAffectedParentSize {
         @MockBox {
           size: pipe!(*$child2_size),
           on_performed_layout: move |_| {